+++ src/buttons/props.js
@@ -2,7 +2,7 @@
 
 import { ZalgoPromise } from 'zalgo-promise/src';
 import { values, uniqueID } from 'belter/src';
-import { createOrder, type OrderCreateRequest,
+import { type OrderCreateRequest,
     type OrderGetResponse, type OrderCaptureResponse, type OrderAuthorizeResponse } from '@paypal/sdk-client/src';
 import { FUNDING, PLATFORM, INTENT, COMMIT, VAULT,
     ENV, COUNTRY, LANG, COUNTRY_LANGS, type LocaleType, CARD, COMPONENTS } from '@paypal/sdk-constants/src';
@@ -12,7 +12,7 @@ import { BUTTON_LABEL, BUTTON_COLOR, BUTTON_LAYOUT, BUTTON_SHAPE, BUTTON_SIZE } 
 import { getFundingConfig } from '../funding';
 import type { FundingEligibilityType } from '../types';
 
-import { BUTTON_SIZE_STYLE } from './config';
+import { BUTTON_SIZE_STYLE } from './template/config';
 
 export type CreateOrderData = {|
 
@@ -34,12 +34,30 @@ export type OnApproveData = {|
 
 export type CreateBillingAgreement = () => ZalgoPromise<string> | string;
 
+export type CreateSubscriptionRequest = {||};
+export type SubscriptionResponse = {||} | {};
+export type CreateSubscriptionData = {|
+|} | {};
+
+export type CreateSubscriptionActions = {|
+    subscription : {
+        create : (CreateSubscriptionRequest) => ZalgoPromise<string>,
+        revise : (CreateSubscriptionRequest) => ZalgoPromise<string>
+    }
+|};
+
+export type CreateSubscription = (CreateSubscriptionData, CreateSubscriptionActions) => ZalgoPromise<string> | string;
+
 export type OnApproveActions = {|
     redirect : (string, CrossDomainWindowType) => ZalgoPromise<void>,
     order : {
         capture : () => ZalgoPromise<OrderCaptureResponse>,
         get : () => ZalgoPromise<OrderGetResponse>,
         authorize : () => ZalgoPromise<OrderAuthorizeResponse>
+    },
+    subscription : {
+        get : () => ZalgoPromise<SubscriptionResponse>,
+        activate : () => ZalgoPromise<SubscriptionResponse>
     }
 |};
 
@@ -134,7 +152,8 @@ export type RenderButtonProps = {|
     sessionID : string,
     buttonSessionID : string,
     nonce : string,
-    components : $ReadOnlyArray<$Values<typeof COMPONENTS>>
+    components : $ReadOnlyArray<$Values<typeof COMPONENTS>>,
+    onShippingChange : ?Function
 |};
 
 export type PrerenderDetails = {|
@@ -145,19 +164,15 @@ export type PrerenderDetails = {|
 
 export type GetPrerenderDetails = () => PrerenderDetails | void;
 
-export type ProxyRest = {|
-    createOrder : typeof createOrder
-|};
-
 export type ButtonProps = {|
     intent : $Values<typeof INTENT>,
     createOrder : CreateOrder,
     createBillingAgreement : CreateBillingAgreement,
+    createSubscription : CreateSubscription,
     oncancel : OnCancel,
     onApprove : OnApprove,
     onClick : OnClick,
     getPrerenderDetails : GetPrerenderDetails,
-    proxyRest : ProxyRest,
     style : ButtonStyle,
     locale : LocaleType,
     commit : boolean,
@@ -167,11 +182,12 @@ export type ButtonProps = {|
     platform : $Values<typeof PLATFORM>,
     fundingEligibility : FundingEligibilityType,
     remembered : $ReadOnlyArray<$Values<typeof FUNDING>>,
+    remember : ($ReadOnlyArray<$Values<typeof FUNDING>>) => void,
     clientID : string,
     sessionID : string,
     buttonSessionID : string,
-    nonce : string,
-    proxyRest : ({ createOrder : typeof createOrder }) => ZalgoPromise<void>
+    onShippingChange : ?Function,
+    nonce : string
 |};
 
 export type ButtonPropsInputs = {|
@@ -186,10 +202,12 @@ export type ButtonPropsInputs = {|
     platform? : $PropertyType<ButtonProps, 'platform'> | void,
     fundingEligibility? : $PropertyType<ButtonProps, 'fundingEligibility'> | void,
     remembered? : $PropertyType<ButtonProps, 'remembered'> | void,
+    remember? : $PropertyType<ButtonProps, 'remember'> | void,
     sessionID? : $PropertyType<ButtonProps, 'sessionID'> | void,
     buttonSessionID? : $PropertyType<ButtonProps, 'buttonSessionID'> | void,
     nonce? : string,
     components : $ReadOnlyArray<$Values<typeof COMPONENTS>>,
+    onShippingChange : ?Function,
     csp? : {
         nonce? : string
     }
@@ -302,7 +320,8 @@ export function normalizeButtonProps(props : ?ButtonPropsInputs) : RenderButtonP
         buttonSessionID = uniqueID(),
         csp = {},
         components = [ COMPONENTS.BUTTONS ],
-        nonce = ''
+        nonce = '',
+        onShippingChange
     } = props;
 
     const { country, lang } = locale;
@@ -337,5 +356,5 @@ export function normalizeButtonProps(props : ?ButtonPropsInputs) : RenderButtonP
 
     style = normalizeButtonStyle(style);
 
-    return { clientID, style, locale, remembered, env, fundingEligibility, platform, buttonSessionID, commit, sessionID, nonce, components };
+    return { clientID, style, locale, remembered, env, fundingEligibility, platform, buttonSessionID, commit, sessionID, nonce, components, onShippingChange };
 }