+++ src/checkout/template/index.js
@@ -1,4 +1,3 @@
 /* @flow */
 
 export * from './componentTemplate';
-export * from './containerTemplate';