+++ src/constants/index.js
@@ -1,6 +1,5 @@
 /* @flow */
 
 export * from './button';
-export * from './fpti';
 export * from './misc';
 export * from './class';