+++ src/constants/misc.js
@@ -1,12 +1,5 @@
 /* @flow */
 
-export const LOG_LEVEL = {
-    DEBUG: ('debug' : 'debug'),
-    INFO:  ('info' : 'info'),
-    WARN:  ('warn' : 'warn'),
-    ERROR: ('error' : 'error')
-};
-
 export const ATTRIBUTE = {
     BUTTON:            ('data-button' : 'data-button'),
     FUNDING_SOURCE:    ('data-funding-source' : 'data-funding-source'),
@@ -15,5 +8,3 @@ export const ATTRIBUTE = {
 };
 
 export const DEFAULT = ('default' : 'default');
-
-