+++ src/funding/bancontact/config.jsx
@@ -11,6 +11,8 @@ export function getBancontactConfig() : FundingSourceConfig {
         ...DEFAULT_FUNDING_CONFIG,
         
         defaultLabel: BUTTON_LABEL.BANCONTACT,
+
+        shippingChange: false,
     
         layouts: [
             BUTTON_LAYOUT.VERTICAL
@@ -20,7 +22,7 @@ export function getBancontactConfig() : FundingSourceConfig {
             [ BUTTON_LABEL.BANCONTACT ]: {
                 ...DEFAULT_LABEL_CONFIG,
     
-                Label: BancontactLogo
+                Label: ({ logoColor, optional }) => BancontactLogo({ logoColor, optional })
             }
         }
     };