+++ src/funding/wechatpay/config.jsx
@@ -11,6 +11,8 @@ export function getWechatpayConfig() : FundingSourceConfig {
         ...DEFAULT_FUNDING_CONFIG,
 
         defaultLabel: BUTTON_LABEL.WECHATPAY,
+
+        shippingChange: false,
     
         layouts: [
             BUTTON_LAYOUT.VERTICAL
@@ -20,7 +22,7 @@ export function getWechatpayConfig() : FundingSourceConfig {
             [ BUTTON_LABEL.WECHATPAY ]: {
                 ...DEFAULT_LABEL_CONFIG,
     
-                Label: WechatpayLogo
+                Label: ({ logoColor, optional }) => WechatpayLogo({ logoColor, optional })
             }
         }
     };